renamed motion_graph to motion_statechart
