+++ giskardpy/goals/align_to_push_door.py
@@ -6,7 +6,7 @@ from giskardpy.data_types.data_types import ColorRGBA
 from giskardpy.data_types.exceptions import GoalInitalizationException
 from giskardpy.goals.goal import Goal
 from giskardpy.god_map import god_map
-from giskardpy.motion_graph.tasks.task import WEIGHT_BELOW_CA, Task
+from giskardpy.motion_statechart.tasks.task import WEIGHT_BELOW_CA, Task
 
 
 class AlignToPushDoor(Goal):
