+++ giskardpy/motion_statechart/graph_node.py
