+++ giskardpy/motion_statechart/monitors/cartesian_monitors.py
@@ -3,7 +3,7 @@ from typing import Optional
 import giskardpy.casadi_wrapper as cas
 from giskardpy.data_types.data_types import PrefixName
 from giskardpy.god_map import god_map
-from giskardpy.motion_graph.monitors.monitors import Monitor
+from giskardpy.motion_statechart.monitors.monitors import Monitor
 
 
 class PoseReached(Monitor):
