+++ giskardpy/motion_statechart/monitors/joint_monitors.py
@@ -1,7 +1,7 @@
 from typing import Dict, Optional
 
 import giskardpy.casadi_wrapper as cas
-from giskardpy.motion_graph.monitors.monitors import Monitor
+from giskardpy.motion_statechart.monitors.monitors import Monitor
 from giskardpy.god_map import god_map
 from giskardpy.data_types.data_types import Derivatives
