+++ giskardpy/motion_statechart/monitors/set_prediction_horizon.py
@@ -5,7 +5,7 @@ from giskardpy.qp.qp_solver_ids import SupportedQPSolver
 import giskardpy.casadi_wrapper as cas
 from giskardpy.data_types.exceptions import MonitorInitalizationException
 from giskardpy.god_map import god_map
-from giskardpy.motion_graph.monitors.monitors import PayloadMonitor
+from giskardpy.motion_statechart.monitors.monitors import PayloadMonitor
 
 
 class SetPredictionHorizon(PayloadMonitor):
