+++ giskardpy/motion_statechart/tasks/align_planes.py
@@ -3,7 +3,7 @@ from typing import Optional
 import giskardpy.casadi_wrapper as cas
 from giskardpy.data_types.data_types import ColorRGBA, PrefixName
 from giskardpy.goals.goal import Goal
-from giskardpy.motion_graph.tasks.task import WEIGHT_ABOVE_CA, Task
+from giskardpy.motion_statechart.tasks.task import WEIGHT_ABOVE_CA, Task
 from giskardpy.middleware import get_middleware
 from giskardpy.god_map import god_map
