+++ giskardpy/motion_statechart/tasks/cartesian_tasks.py
@@ -4,8 +4,8 @@ from giskardpy.data_types.data_types import Derivatives
 from giskardpy import casadi_wrapper as cas
 from giskardpy.data_types.data_types import PrefixName, ColorRGBA
 from giskardpy.god_map import god_map
-from giskardpy.motion_graph.monitors.cartesian_monitors import PositionReached, OrientationReached
-from giskardpy.motion_graph.tasks.task import Task, WEIGHT_ABOVE_CA
+from giskardpy.motion_statechart.monitors.cartesian_monitors import PositionReached, OrientationReached
+from giskardpy.motion_statechart.tasks.task import Task, WEIGHT_ABOVE_CA
 from giskardpy.symbol_manager import symbol_manager
