+++ giskardpy/motion_statechart/tasks/feature_functions.py
@@ -5,7 +5,7 @@ from typing import Optional, Union
 from giskardpy import casadi_wrapper as cas
 from giskardpy.data_types.data_types import PrefixName, ColorRGBA
 from giskardpy.god_map import god_map
-from giskardpy.motion_graph.tasks.task import WEIGHT_BELOW_CA, Task
+from giskardpy.motion_statechart.tasks.task import WEIGHT_BELOW_CA, Task
 
 
 class FeatureFunctionGoal(Task):
